adjust data gens.
