+++ scripts/data_generators/DynamicPosesDataSaver.py
@@ -15,8 +15,8 @@ from std_msgs.msg import Float32
 
 sys.path.append(rospkg.RosPack().get_path('ros_robotic_skin'))
 from scripts import utils  # noqa: E402
-from scripts.controllers.PandaController import PandaController  # noqa: E402
-from scripts.controllers.SawyerController import SawyerController  # noqa: E402
+from scipts.controllers.RobotController import PandaController, SawyerController  # noqa: E402
+
 
 RAD2DEG = 180.0 / np.pi
 OSCILLATION_TIME = 3.0
@@ -266,6 +266,9 @@ class DynamicPoseDataSaver():
         for i in range(total_imu_topics):
             self.imu_names.append('imu_link{}'.format(i))
             self.imu_topics.append('imu_data{}'.format(i))
+        """
+        to-do get the joints the imus are connected to.
+        """
         self.ready = False
         self.curr_positions = [0, 0, 0, 0, 0, 0, 0]
         self.curr_pose_name = self.pose_names[0]
